+++ distro/vhx-quartz.select.js
@@ -137,9 +137,18 @@ vhxm.components.shared.select.ui.container = {
     return new vhxm.components.shared.select.controller(opts);
   },
   view: function view(ctrl, opts) {
+    opts.custom_trigger ? opts.custom_trigger.attrs.onclick = ctrl.handleClick : '';
+    opts.custom_trigger ? opts.custom_trigger.attrs.className += ' c-select--trigger' : '';
+
     return m('.c-select--container.form' + (opts.search ? '.has-search' : '') + (opts.inline ? '.inline' : ''), {
+      config: function config(el, isInitialized) {
+        if (opts.custom_trigger && isInitialized) {
+          var left_pos = el.querySelector('.c-select--trigger').offsetWidth * .25;
+          el.querySelector('.c-select--caret').style.left = left_pos + 'px';
+        }
+      },
       onkeydown: ctrl.handleKeydown
-    }, [m('a.c-select--trigger.btn-dropdown-' + (opts.color ? opts.color : 'gray') + '.btn--fill' + (ctrl.state.isDropdownOpen() ? '.is-active' : ''), {
+    }, [opts.custom_trigger ? opts.custom_trigger : m('a.c-select--trigger.btn-dropdown-' + (opts.color ? opts.color : 'gray') + '.btn--fill' + (ctrl.state.isDropdownOpen() ? '.is-active' : ''), {
       href: '#',
       onclick: ctrl.handleClick
     }, ctrl.selectedLabel()), m('.c-select--dropdown' + (ctrl.state.isDropdownOpen() ? '.is-open' : ''), [opts.search ?
@@ -160,7 +169,7 @@ vhxm.components.shared.select.ui.container = {
       type: 'hidden',
       name: opts.name,
       value: ctrl.state.selected().value
-    })])]);
+    }), m('span.c-select--caret')])]);
   }
 };