+++ quartz-js/components/radio/radio.state.js
@@ -1,0 +1,3 @@
+vhxm.components.shared.radio.state = function() {
+  this.isChecked = m.prop(null);
+};