+++ quartz-js/components/scope.js
@@ -1,5 +1,8 @@
 if (typeof(vhxm) !== 'undefined') {
   vhxm.components.shared = {
+    modal: {
+      ui: {}
+    },
     radio: {
       ui: {}
     },
@@ -37,5 +40,8 @@ if (typeof(vhxm) !== 'undefined') {
     },
     avatar: {},
     size: {},
+    tag: {
+      ui: {}
+    }
   };
 }