+++ quartz-js/components/select/docs/guide.html.js
@@ -13,6 +13,7 @@ Q.components.guide.js.ui.select.container = {
           view: function() {
             return m('div', [
               m.component(vhxm.components.shared.select.ui.container, {
+                custom_trigger: m('a', { href: '#', }, '+ Add Video'),
                 search: function(query, callback) {
                   callback([
                     {
@@ -71,7 +72,8 @@ Q.components.guide.js.ui.select.container = {
           { name: 'search', type: 'boolean', default_value: 'false', description: 'Toggles inclusion of search input' },
           { name: 'descriptor_prop', type: 'string', default_value: 'null', description: 'Supplemental item key to look for in the provided items to display a descriptor for each item.' },
           { name: 'item_type', type: 'string', default_value: 'standard', description: 'standard OR media - sets UI of item in dropdown list' },
-          { name: 'footer_action', type: 'boolean', default_value: 'false', description: 'Currently for Collections - show a Create New Category action at the bottom of dropdown.' }
+          { name: 'footer_action', type: 'boolean', default_value: 'false', description: 'Currently for Collections - show a Create New Category action at the bottom of dropdown.' },
+          { name: 'custom_trigger', type: 'mithril element', default_value: 'null', required: false, description: 'Trigger element that will replace default button dropdown' }
         ]
       })
     ]);