+++ quartz-js/components/select/select.ui.js
@@ -4,9 +4,20 @@ vhxm.components.shared.select.ui.container = {
     return new vhxm.components.shared.select.controller(opts);
   },
   view: function(ctrl, opts) {
+    opts.custom_trigger ? opts.custom_trigger.attrs.onclick = ctrl.handleClick : '';
+    opts.custom_trigger ? opts.custom_trigger.attrs.className += ' c-select--trigger' : '';
+
     return m('.c-select--container.form' + (opts.search ? '.has-search' : '') + (opts.inline ? '.inline' : ''), {
+        config: function(el, isInitialized) {
+          if (opts.custom_trigger && isInitialized) {
+            var left_pos = el.querySelector('.c-select--trigger').offsetWidth * .25;
+            el.querySelector('.c-select--caret').style.left = left_pos + 'px';
+          }
+        },
         onkeydown: ctrl.handleKeydown
       }, [
+      opts.custom_trigger ?
+      opts.custom_trigger :
       m('a.c-select--trigger.btn-dropdown-' + (opts.color ? opts.color : 'gray') + '.btn--fill' + (ctrl.state.isDropdownOpen() ? '.is-active' : ''), {
         href: '#',
         onclick: ctrl.handleClick
@@ -36,7 +47,8 @@ vhxm.components.shared.select.ui.container = {
           m('a.c-select--footer-link.block.text--bold.text-center.padding-vert-small', {
             href: '#'
           }, opts.footer_action + (ctrl.state.searchInputValue().length ? (' \'' + ctrl.state.searchInputValue() + '\'') : ''))
-        ]) : ''
+        ]) : '',
+        m('span.c-select--caret')
       ])
     ]);
   }