+++ quartz-js/components/sidebar/sidebar.state.js
@@ -3,6 +3,7 @@ vhxm.components.shared.sidebar.state = {
   isLoaded       : m.prop(false),
   skipTransition : m.prop(false),
   template       : m.prop(null),
+  onBeforeClose  : function() {},
   onClose        : function() {},
   onOpen         : function() {}
 };