+++ quartz-js/components/tag/docs/code/tag.js
@@ -1,0 +1,9 @@
+m.component(vhxm.components.shared.tag.ui.container, {
+  label: 'Comedy',
+  onShow: function() {
+    window.alert('show something');
+  },
+  onRemove: function() {
+    window.alert('remove something');
+  }
+});