+++ quartz-js/components/tag/tag.controller.js
@@ -1,0 +1,5 @@
+vhxm.components.shared.tag.controller = function(opts) {
+  let self = this;
+
+  self.state = new vhxm.components.shared.tag.state();
+};